refactor: split User and Auth modules for better separation of concerns
