+++ src/apps/auth/api/auth.controller.ts
@@ -64,14 +64,5 @@ export class AuthController extends BaseController {
     }
   }
 
-  async getAllUsers(req: Request, res: Response, next: NextFunction): Promise<void> {
-    try {
-      const authService = new AuthService("basic"); 
-      const users = await authService.getAllUsers(); 
-      this.logger.info(`Fetched ${users.length} users`); 
-      ApiResponse.success(res, "Users fetched successfully", users); 
-    } catch (error: any) {
-      next(error); 
-    }
-  }
+ 
 }
