+++ src/apps/auth/data-access/user.repository.ts
@@ -15,9 +15,4 @@ export class UserRepository extends MongooseRepository<IUser> {
     return user;
   }
 
-  
-  async getAllUsers(): Promise<Omit<IUser, "password">[]> {
-    const users = await this.model.find({}, { password: 0 }).exec(); 
-    return users.map((user) => user.toObject() as Omit<IUser, "password">);
-  }
 }
