+++ src/index.ts
@@ -6,6 +6,7 @@ import { swaggerDocs } from "./config/swagger/swagger";
 import { LoggerConfig } from "./config/logger/logger.config";
 import { MongooseConfig } from "./config/mongoose/mongoose.config";
 import authRoute from "./apps/auth/api/auth.route";
+import userRoute from "./apps/auth/api/user.route";
 import { corsMiddleware } from "./librairies/middlewares/cors.middleware";
 import { rateLimiterMiddleware } from "./librairies/middlewares/rate.limit.middleware";
 import { errorHandlerMiddleware } from "./librairies/middlewares/error.middleware";
@@ -33,6 +34,9 @@ async function startApp() {
 
     // Authentication routes
     app.use("/api/auth", authRoute(router));
+
+    // User routes
+    app.use("/api/users", userRoute(router));
 
     // Error handling middleware
     app.use(errorHandlerMiddleware as express.ErrorRequestHandler);
